+++ NEWS.md
@@ -4,7 +4,8 @@
  * Added classification settings loading for mapping
  * Added instance-format settings loading for mapping
  * Added contributor name types setting loading for mapping
-
+ * Added contributor types setting loading for mapping, simple mapping for primary sign and name
+ 
 ## 2019-09-09 v1.6.0
  * Progress mechanism was updated
  * Changed RawRecordsDto schema with extended metadata information